% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.interactome.hs.R
\docType{data}
\name{annotation.interactome.hs}
\alias{annotation.interactome.hs}
\title{Annotation of human interactome}
\format{
An object of class \code{list} of length 15.
}
\usage{
annotation.interactome.hs
}
\description{
Annotation of each category. Used to split the CoPPI algorithm on
    on each category
}
\keyword{datasets}
