% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_and_mapping.R
\name{annotation_and_mapping}
\alias{annotation_and_mapping}
\title{annotation_and_mapping}
\usage{
annotation_and_mapping(all.interacting.proteins, tax_id, AnnNbyN)
}
\arguments{
\item{all.interacting.proteins}{vector of proteins that need to be annotated}

\item{tax_id}{taxonomy id of the species}

\item{AnnNbyN}{number to split 'all.interacting.proteins' to retrieve annotation
on StringDB}
}
\value{
A list with mapped ids data.frame and the annotation data.frame
}
\description{
annotation_and_mapping
}
