% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_interactome.R
\name{build_interactome}
\alias{build_interactome}
\title{build_interactome}
\usage{
build_interactome(
  directory_interactome,
  tax_id,
  interactome = NULL,
  scores_threshold = NULL,
  AnnNbyN = 1999
)
}
\arguments{
\item{directory_interactome}{directory of String interactome file. ignored if
interactome is not NULL}

\item{tax_id}{taxonomy id of the specie}

\item{interactome}{interactome data.frame, default = NULL.}

\item{scores_threshold}{thresholds vector, if NULL no filtering will be applied}

\item{AnnNbyN}{vector to split the searching on String. If the function gives
Timeout Error try to lower the value (e.g 999 or 1499). Default = 1999.}
}
\value{
List containing information about interactome
}
\description{
build_interactome
}
