% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.hs.R
\docType{data}
\name{categories.hs}
\alias{categories.hs}
\title{Categories of annotations}
\format{
An object of class \code{character} of length 14.
}
\usage{
categories.hs
}
\description{
Human categories of annotation avaliables to perform CoPPIs.
}
\keyword{datasets}
