% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.interactome.hs.R
\docType{data}
\name{categories.interactome.hs}
\alias{categories.interactome.hs}
\title{Categories of human interactome}
\format{
An object of class \code{list} of length 15.
}
\usage{
categories.interactome.hs
}
\description{
Interactome of each category. Used to lower the necessary proteins
    and lower the influence of pvalue correction.
}
\keyword{datasets}
