% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_annotation.R
\name{compute_annotation}
\alias{compute_annotation}
\title{Title}
\usage{
compute_annotation(genes_id, annotation.df)
}
\arguments{
\item{genes_id}{vector containing genes ids}

\item{annotation.df}{special data.frame containing annotation}
}
\value{
annotation data.frame with just and only the biological terms involved in the genes id list
}
\description{
Title
}
