% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPPI.R
\name{CoPPI}
\alias{CoPPI}
\title{Title}
\usage{
CoPPI(
  annotations,
  correlations,
  perc.ned.nppi,
  perc.sign,
  significance.coppi,
  correction.coppi,
  min_edges,
  max_edges,
  names_of_groups,
  p.filter,
  score.filter,
  interactome
)
}
\arguments{
\item{annotations}{special annotation object}

\item{correlations}{special correlation object}

\item{perc.ned.nppi}{filtering threshold, minimum percentage of significant number of edges divided by number of PPIs}

\item{perc.sign}{filtering threshold, minimum percentage of significant number of edges on total number of edges}

\item{significance.coppi}{significance threshold for CoPPIs terms pvalues}

\item{correction.coppi}{correction method for CoPPIs terms}

\item{min_edges}{filtering threshold, minimum edges of a biological term}

\item{max_edges}{filtering threshold, maximum edges of a biological term}

\item{names_of_groups}{names of the groups}

\item{p.filter}{filtering threshold, maximum CoPPIs pvalue}

\item{score.filter}{filtering threshold, minimum CoPPIs score}

\item{interactome}{interactome as data.frame}
}
\value{
List with information about CoPPIs result
}
\description{
Title
}
