% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPPIs_pipeline.R
\name{CoPPIs_pipeline}
\alias{CoPPIs_pipeline}
\title{CoPPIs_pipeline}
\usage{
CoPPIs_pipeline(
  dataset,
  genes_id,
  names_of_groups,
  name_analysis,
  interactome = interactome.hs,
  annotation.interactome = annotation.interactome.hs,
  categories.interactome = categories.interactome.hs,
  p.filter = 0.001,
  score.filter = 2,
  gCOR.groups.l = NULL,
  gCOR.categories = NULL,
  signifCorr = 0.05,
  correctionCorr = "BH",
  min_n.corr = NULL,
  significance.coppi = 0.05,
  perc.ned.nppi = 0.01,
  perc.sign = 0.3,
  correction.coppi = "BH",
  corr.test = "spearman",
  compute_weights = T,
  min_edges = 1,
  max_edges = Inf,
  categories = c("Component", "Process", "RCTM", "WikiPathways"),
  threshold_edge_cyto = 0.6,
  Cytoscape = T
)
}
\arguments{
\item{dataset}{The dataset ProteinxSamples}

\item{genes_id}{genes id vector}

\item{names_of_groups}{names of the groups}

\item{name_analysis}{name of the analysis to be saved}

\item{interactome}{interactome}

\item{annotation.interactome}{annotation of protein interactions}

\item{categories.interactome}{interactome for each category}

\item{p.filter}{threshold CoPPIs pvalue, default 0.001}

\item{score.filter}{threshold CoPPIs score, default 2}

\item{gCOR.groups.l}{if a precedent analyisis was run, it could be used to speed up the computational time}

\item{gCOR.categories}{if a precedent analyisis was run, it could be used to speed up the computational time}

\item{signifCorr}{threshold significance correlation, default 0.05}

\item{correctionCorr}{correction method for pvalue correlation, default BH}

\item{min_n.corr}{threshold minimum observation pair for correlation, default NULL}

\item{significance.coppi}{threshold significance for CoPPIs pvalue, default 0.05}

\item{perc.ned.nppi}{threshold percentage number edges number ppi, default 0.01}

\item{perc.sign}{threshold percentage number edges significants on total edges, default 0.3}

\item{correction.coppi}{method for correction pvalues CoPPI, default BH}

\item{corr.test}{correlation test, pearson or spearman, default Spearman}

\item{compute_weights}{if T transformation formula will be applied to the non significant correlation according to their pvalues, default T}

\item{min_edges}{threshold, minimum edges of biological terms, default 1}

\item{max_edges}{threshold, maximum edges of biological terms, default Inf}

\item{categories}{name of the biological categories on to which apply CoPPIs, default  c("Component","Process","RCTM","WikiPathways")}

\item{threshold_edge_cyto}{threshold of similarity edges between terms to be loaded on Cytoscape, default 0.6}

\item{Cytoscape}{if T the result will be loaded on Cytoscape, default T}
}
\value{
List file with every information about CoPPIs output
}
\description{
CoPPIs_pipeline
}
