% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_time_difference.R
\name{format_time_difference}
\alias{format_time_difference}
\title{Title}
\usage{
format_time_difference(time_diff)
}
\arguments{
\item{time_diff}{function that rewrites the time difference}
}
\value{
A string
}
\description{
Title
}
