% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funscore.R
\name{funscore}
\alias{funscore}
\title{Title}
\usage{
funscore(x, y)
}
\arguments{
\item{x}{Number of signficative edges of a biological term}

\item{y}{Number of PPIs of the same biological term}
}
\value{
The CoPPIs score value
}
\description{
Title
}
