% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi_interactions.R
\name{get_ppi_interactions}
\alias{get_ppi_interactions}
\title{get_ppi_interactions}
\usage{
get_ppi_interactions(interactome, annotation.df, map_id.merged)
}
\arguments{
\item{interactome}{interactome object}

\item{annotation.df}{annotation dataframe}

\item{map_id.merged}{mapped id dataframe}
}
\value{
A information list for interactome
}
\description{
get_ppi_interactions
}
