% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_terms.R
\name{graph_terms}
\alias{graph_terms}
\title{graph terms}
\usage{
graph_terms(W, S)
}
\arguments{
\item{W}{Weights}

\item{S}{Scores}
}
\value{
List information about terms similarity network
}
\description{
graph terms
}
