% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome.hs.R
\docType{data}
\name{interactome.hs}
\alias{interactome.hs}
\title{Human Interactome}
\format{
An object of class \code{data.frame} with 2065264 rows and 12 columns.
}
\usage{
interactome.hs
}
\description{
Human Interactome in data.frame format. columns are protein names
    and various scores
}
\keyword{datasets}
