% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load2Cytoscape.R
\name{Load2Cytoscape}
\alias{Load2Cytoscape}
\title{Title}
\usage{
Load2Cytoscape(file_path, filter_similarity = 0.6)
}
\arguments{
\item{file_path}{directory path in which the CoPPIs output excels are}

\item{filter_similarity}{filtering threshold, minimum similarity score between terms}
}
\value{
message of loading states
}
\description{
Title
}
