% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_abs_cor.R
\name{mean_abs_cor}
\alias{mean_abs_cor}
\title{Title}
\usage{
mean_abs_cor(Graph_pathways.groups, g1)
}
\arguments{
\item{Graph_pathways.groups}{graph pathways object}

\item{g1}{group selected}
}
\value{
a data.frame
}
\description{
Title
}
