% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_annotations.R
\name{merge_annotations}
\alias{merge_annotations}
\title{merge_annotations}
\usage{
merge_annotations(df1, df2)
}
\arguments{
\item{df1}{first annotation data.frame}

\item{df2}{second annotation data.frame}
}
\value{
merged annotation data.frame
}
\description{
merge_annotations
}
