% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparison.R
\name{pairwise_comparison}
\alias{pairwise_comparison}
\title{Title}
\usage{
pairwise_comparison(
  tab_k,
  g1,
  g2,
  sd_groups,
  tab_ppicor,
  perc.sign.edges.f,
  perc.sign,
  significance.coppi,
  correction.coppi
)
}
\arguments{
\item{tab_k}{matrix of k ratios}

\item{g1}{index of group 1}

\item{g2}{index of group 2}

\item{sd_groups}{standard deviation of correlations}

\item{tab_ppicor}{table with PPI and correlation associated}

\item{perc.sign.edges.f}{table with percentage of significant edges on total edges}

\item{perc.sign}{filtering threshold, percentage of significant edges on total edges}

\item{significance.coppi}{significance thresold for CoPPIs pvalue}

\item{correction.coppi}{correction method for CoPPIs pvalues}
}
\value{
A score data.frame
}
\description{
Title
}
