% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Title}
\usage{
save_results(category, name_analysis)
}
\arguments{
\item{category}{category object of CoPPIs output}

\item{name_analysis}{name of the analysis to be saved}
}
\value{
message of saving state
}
\description{
Title
}
