% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCytoStyle.R
\name{setCytoStyle}
\alias{setCytoStyle}
\title{Title}
\usage{
setCytoStyle(
  SUID_list = getNetworkList(getSUIDs = T),
  style.name = "score",
  palette.nodes = c("#4EB3D3", "#D4FFD0", "#FFE0FF", "#FC0000")
)
}
\arguments{
\item{SUID_list}{vector of network SUIDs, default getNetworkList(getSUIDs = T)}

\item{style.name}{name of the style, default "score"}

\item{palette.nodes}{palette of the nodes, default c('#4EB3D3','#D4FFD0','#FFE0FF','#FC0000')}
}
\description{
Title
}
