% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{summaryTable}
\usage{
summaryTable(list_directories, rm_part_categ, rm_part.wb, name_analysis)
}
\arguments{
\item{list_directories}{list of directories that contains coppis output}

\item{rm_part_categ}{character vector taht contain eventual part to remove}

\item{rm_part.wb}{character to remove from worksheet}

\item{name_analysis}{name to save the summary}
}
\value{
0 Value if all correct
}
\description{
summaryTable
}
